moved plot_format(), plot_encut_conv(), calculate_encut_conv(), plot_kpoint_conv(), and calculate_kpoint_conv() to aggregate_extraction.py
